refactor(ui): TMS-2508 tenant list renames
Moved the tenant list and card components into the tenant directory. Renamed tenant card to tenant list card to make it clearer where it's used.
